- renamed some folders and file for cleanliness - renamed resource objects to pre-empt conflicts - renamed step objects to pre-empt conflicts - fixed 0-dollar fees bug - misc cleanup
