--- src/resources/Dissolutions/index.ts
@@ -1,8 +1,0 @@
-export { BenefitCompanyDissolutionResource } from '@/resources/Dissolutions/benefitCompany'
-export { CommunityContributionCompanyDissolutionResource } from '@/resources/Dissolutions/communityContributionCompany'
-export { CooperativeDissolutionResource } from '@/resources/Dissolutions/cooperative'
-export { LimitedCompanyDissolutionResource } from '@/resources/Dissolutions/limitedCompany'
-export { UnlimitedDissolutionResource } from '@/resources/Dissolutions/unlimitedCompany'
-
-export { SoleProprietorshipDissolutionResource } from '@/resources/Dissolutions/soleProprietorship'
-export { GeneralPartnershipDissolutionResource } from '@/resources/Dissolutions/generalPartnership'
