+++ src/resources/Incorporation/BC.ts
@@ -1,15 +1,15 @@
 import { IncorporationResourceIF } from '@/interfaces'
 import { CorpTypeCd, FilingCodes, RuleIds } from '@/enums'
-import { BaseStepsTemplate } from './stepTemplates'
+import { IncorporationStepsCorp } from './steps'
 import { GetCorpFullDescription } from '@bcrs-shared-components/corp-type-module'
 
-export const UnlimitedCompanyResource: IncorporationResourceIF = {
-  entityType: CorpTypeCd.BC_ULC_COMPANY,
-  displayName: GetCorpFullDescription(CorpTypeCd.BC_ULC_COMPANY),
-  steps: BaseStepsTemplate,
+export const IncorporationResourceBc: IncorporationResourceIF = {
+  entityType: CorpTypeCd.BC_COMPANY,
+  displayName: GetCorpFullDescription(CorpTypeCd.BC_COMPANY),
+  steps: IncorporationStepsCorp,
   filingData: [{
-    entityType: CorpTypeCd.BC_ULC_COMPANY,
-    filingTypeCode: FilingCodes.INCORPORATION_ULC
+    entityType: CorpTypeCd.BC_COMPANY,
+    filingTypeCode: FilingCodes.INCORPORATION_BC
   }],
   peopleAndRoles: {
     header: '1. Add People or Corporations/Firms to your Application',
@@ -40,15 +40,9 @@ export const UnlimitedCompanyResource: IncorporationResourceIF = {
     countMinimum: 1
   },
   incorporationArticles: {
-    articles: 'Unlimited Liability Company Articles',
-    articlesTooltip: 'The Articles for an Unlimited Liability Company must include the following statement: “The ' +
-    'shareholders of this company are jointly and severally liable to satisfy the debts and liabilities of this ' +
-    'company to the extent provided in section 51.3 of the Business Corporations Act.” Articles should also outline ' +
-    'the rules and procedures for corporate matters such as holding meetings, issuing and transferring shares, and ' +
-    'duties of directors and officers.',
-    provisions: 'liability provision',
-    provisionTooltip: 'A clause in the Articles which communicates the liability statement required by the Business ' +
-    'Corporations Act 51.11.'
+    articles: '<br>BC LTD Articles.', // Line break is required to prevent tooltip from splitting the screen.
+    articlesTooltip: 'Articles should outline the rules and procedures for corporate matters such as holding ' +
+    'meetings, issuing and transferring shares, and duties of directors and officers.'
   },
   incorporationAgreement: {
     helpSection: [
@@ -59,26 +53,22 @@ export const UnlimitedCompanyResource: IncorporationResourceIF = {
             incorporation agreement and articles for your company. It uses all the standard provisions by legislation. 
             There are three types of samples depending on if you're incorporating a Limited Company, a Benefit Company, 
             or a Community Contribution Company.`,
-          `If you would like to customize any other provisions in the Articles, you cannot use these samples.
+          `If you would like to customize any other provisions in the Articles, you cannot use these samples. 
             We recommend seeking professional assistance from a lawyer or accountant to help you prepare your articles.`
-        ]
-      },
-      {
-        header: 'What is a Liability Provision?',
-        helpText: [
-          `A Liability Provision is a statement by the company of the shareholders joint liability to satisfy all debts
-            and liabilities of the company.`,
-          'An Unlimited Liability Company must include a Liability Provision in its Articles.'
         ]
       }
     ],
-    article: 'unlimited_liability_company_incorporation_agreement.pdf',
+    article: 'bc_limited_company_incorporation_agreement.pdf',
     documents: [
       {
+        code: 'sample',
+        description: 'The <b>sample Incorporation Agreement and Articles</b> ' +
+          'has been completed and a copy has been added to the company\'s record book.'
+      },
+      {
         code: 'custom',
-        description: 'The <b>custom Incorporation Agreement and custom Articles</b> containing the liability' +
-        ' provision required by the Business Corporations Act 51.11 has been completed and a copy has been added' +
-        ' to the company\'s record book.'
+        description: 'The <b>custom Incorporation Agreement and custom Articles</b> ' +
+          'has been completed and a copy has been added to the company\'s record book.'
       }
     ]
   },
