+++ src/resources/Incorporation/CC.ts
@@ -1,15 +1,15 @@
 import { IncorporationResourceIF } from '@/interfaces'
 import { CorpTypeCd, FilingCodes, RuleIds } from '@/enums'
-import { BaseStepsTemplate } from './stepTemplates'
+import { IncorporationStepsCorp } from './steps'
 import { GetCorpFullDescription } from '@bcrs-shared-components/corp-type-module'
 
-export const LimitedCompanyResource: IncorporationResourceIF = {
-  entityType: CorpTypeCd.BC_COMPANY,
-  displayName: GetCorpFullDescription(CorpTypeCd.BC_COMPANY),
-  steps: BaseStepsTemplate,
+export const IncorporationResourceCc: IncorporationResourceIF = {
+  entityType: CorpTypeCd.BC_CCC,
+  displayName: GetCorpFullDescription(CorpTypeCd.BC_CCC),
+  steps: IncorporationStepsCorp,
   filingData: [{
-    entityType: CorpTypeCd.BC_COMPANY,
-    filingTypeCode: FilingCodes.INCORPORATION_BC
+    entityType: CorpTypeCd.BC_CCC,
+    filingTypeCode: FilingCodes.INCORPORATION_CC
   }],
   peopleAndRoles: {
     header: '1. Add People or Corporations/Firms to your Application',
@@ -31,8 +31,8 @@ export const LimitedCompanyResource: IncorporationResourceIF = {
       },
       {
         id: RuleIds.NUM_DIRECTORS,
-        text: 'At least one Director',
-        test: (num) => { return (num >= 1) }
+        text: 'At least three Directors',
+        test: (num) => { return (num >= 3) }
       }
     ]
   },
@@ -40,9 +40,16 @@ export const LimitedCompanyResource: IncorporationResourceIF = {
     countMinimum: 1
   },
   incorporationArticles: {
-    articles: '<br>BC LTD Articles.', // Line break is required to prevent tooltip from splitting the screen.
-    articlesTooltip: 'Articles should outline the rules and procedures for corporate matters such as holding ' +
-    'meetings, issuing and transferring shares, and duties of directors and officers.'
+    articles: 'Community Contribution Company Articles',
+    articlesTooltip: 'The Articles for a Community Contribution Company must include the following statements “This ' +
+    'company is a community contribution company, and, as such, has purposes beneficial to society. This company is ' +
+    'restricted, in accordance with Part 2.2 of the Business Corporations Act, in its ability to pay dividends and ' +
+    'to distribute its assets on dissolution or otherwise.” Articles should also outline the rules and procedures ' +
+    'for corporate matters such as holding meetings, issuing and transferring shares, and duties of directors ' +
+    'and officers.',
+    provisions: 'community provision',
+    provisionTooltip: 'A clause in the Articles which communicates the liability statement required by the Business ' +
+    'Corporations Act 51.11.'
   },
   incorporationAgreement: {
     helpSection: [
@@ -50,25 +57,29 @@ export const LimitedCompanyResource: IncorporationResourceIF = {
         header: 'What is the sample Incorporation Agreement and Company Articles?',
         helpText: [
           `The sample Incorporation Agreement and Company Articles is a template that you can use to create an 
-            incorporation agreement and articles for your company. It uses all the standard provisions by legislation. 
+            Incorporation Agreement and articles for your company. It uses all the standard provisions by legislation. 
             There are three types of samples depending on if you're incorporating a Limited Company, a Benefit Company, 
-            or a Community Contribution Company.`,
+            Unlimited Liability Company, or a Community Contribution Company.`,
           `If you would like to customize any other provisions in the Articles, you cannot use these samples. 
             We recommend seeking professional assistance from a lawyer or accountant to help you prepare your articles.`
         ]
-      }
-    ],
-    article: 'bc_limited_company_incorporation_agreement.pdf',
-    documents: [
-      {
-        code: 'sample',
-        description: 'The <b>sample Incorporation Agreement and Articles</b> ' +
-          'has been completed and a copy has been added to the company\'s record book.'
       },
       {
+        header: 'What is a Community Provision?',
+        helpText: [
+          `A Community Provision is a statement by the company of its benefits to society and its restrictions in its 
+            ability to pay dividends and distribute assets.`,
+          `A Community Contribution Company must include a community provision in its Articles.`
+        ]
+      }
+    ],
+    article: 'community_contribution_company_incorporation_agreement.pdf',
+    documents: [
+      {
         code: 'custom',
-        description: 'The <b>custom Incorporation Agreement and custom Articles</b> ' +
-          'has been completed and a copy has been added to the company\'s record book.'
+        description: 'The <b>custom Incorporation Agreement and custom Articles</b> containing the community' +
+        ' provision required by the Business Corporations Act 51.911 has been completed and a copy has been added' +
+        ' to the company\'s record book.'
       }
     ]
   },
