+++ src/resources/Registration/index.ts
@@ -1,0 +1,2 @@
+export * from '@/resources/Registration/GP'
+export * from '@/resources/Registration/SP'
