--- src/resources/Registrations/index.ts
@@ -1,2 +1,0 @@
-export { GeneralPartnershipResource } from '@/resources/Registrations/generalPartnership'
-export { SoleProprietorshipResource } from '@/resources/Registrations/soleProprietorship'
