+++ src/router/routes.ts
@@ -168,7 +168,7 @@ export const routes = [
       step: 1,
       requiresAuth: true,
       filingType: FilingTypes.VOLUNTARY_DISSOLUTION,
-      noStepper: true // to hide stepper for this view
+      noStepper: true // hide stepper for this route/filing type
     }
   },
   {
